requirements.txt
langchain
openai
streamlit
huggingface_hub
python-dotenv
flask